requirements.txt
django
djangorestframework-simplejwt
djangorestframework
django-extensions
django-cors-headers
psycopg2-binary